requirements.txt
numpy 
pandas 
python-dotenv
scikit-learn
seaborn
Flask
dill
astropy
statsmodels
scipy
xgboost
catboost
dagshub
mlflow
cassandra-driver

-e .